Add files via upload
Simple Linear Regression Model for Predicting Percentage Score per Hour of Study

Overview
This repository contains a simple machine learning project focused on building a linear regression model to predict the percentage score of students based on the number of hours they study. This project serves as a beginner-friendly example to understand the basics of linear regression and how it can be applied to real-world scenarios.

Dataset
The dataset used for training and testing the model is included in the data directory. It consists of two columns: "Hours" (the number of hours a student studies) and "Percentage" (the corresponding percentage score achieved by the student).